Upgrade Gradle to 8.1.1.
Summary:
Release [patch]

Upgrade Gradle and Gradle wrapper to 8.1.1.

I used

```
./gradlew wrapper --gradle-version 8.1.1
```

to upgrade the wrapper and replaced //distributionUrl// with our JFrog.

The reason for this upgrade is to get the fix for this bug: https://github.com/gradle/gradle/issues/24754,
which makes newer builds of this plugin incompatible with older versions of Gradle.

Test Plan: Run it in TeamCity.

Reviewers: sergei.prodanov, O5 elhub/devxp

Reviewed By: sergei.prodanov, O5 elhub/devxp

JIRA Issues: TD-3657

Differential Revision: https://phabricator.elhub.cloud/D10876